Final version of the game, still need to complete the README.MD and make a report

+++ src/Entities/EnemyShip.cpp
@@ -5,12 +5,12 @@
 #include "EnemyShip.h"
 
 namespace Entities {
-    EnemyShip::EnemyShip(const double &x, const double &y) : Entity(std::make_pair(x,y)) {}
+    EnemyShip::EnemyShip(const float &x, const float &y) : Entity(std::make_pair(x,y)) {}
 
     void EnemyShip::moveLeft() {
         // do check if left side of entity reaches the side of the window, if not do move
         if (getCoords().first - (getEntityWidth()/2) > -4) {
-            std::pair<double, double> newCoords = getCoords();
+            std::pair<float, float> newCoords = getCoords();
             newCoords.first -= 0.01;
             setCoords(newCoords);
             notify(*this, "moveLeftEnemy");
@@ -20,7 +20,7 @@ namespace Entities {
     void EnemyShip::moveRight() {
         // do check if right side of entity reaches the side of the window, if not do move
         if (getCoords().first + (getEntityWidth()/2) < 4) {
-            std::pair<double, double> newCoords = getCoords();
+            std::pair<float, float> newCoords = getCoords();
             newCoords.first += 0.01;
             setCoords(newCoords);
             notify(*this, "moveRightEnemy");
@@ -30,7 +30,7 @@ namespace Entities {
     void EnemyShip::moveDown() {
         // do check if right side of entity reaches the side of the window, if not do move
         if (getCoords().second + (getEntityHeight()/2) > -3) {
-            std::pair<double, double> newCoords = getCoords();
+            std::pair<float, float> newCoords = getCoords();
             newCoords.second -= 0.15;
             setCoords(newCoords);
             notify(*this, "moveDownEnemy");
